feat: MAT-47 임시 타입 정의를 api 폴더에 추가
