+++ src/apis/index.ts
@@ -1,2 +1,5 @@
 export * from './example';
 export * from './exampleById';
+export * from './players';
+export * from './teams';
+export * from './matchLogs';
