+++ src/apis/players.ts
@@ -1,0 +1,10 @@
+export interface StartingPlayer {
+  id: number;
+  name: string;
+  number: number;
+  position: string;
+  profileImageUrl: string;
+  goals: number;
+  assists: number;
+  fouls: number;
+}
